+++ src/lib/forms/radio/Radio.svelte
@@ -2,13 +2,12 @@
   import Label from "../label/Label.svelte";
   import { type RadioProps as Props, radio } from ".";
 
-  let { children, aria_describedby, labelClass, color = "primary", group = $bindable<T>(), value = $bindable<T>(), inputClass, custom = false, disabled, ...restProps }: Props<T> = $props();
+  let { children, aria_describedby, labelClass, color = "primary", group = $bindable<T>(), value = $bindable<T>(), inputClass, custom = false, disabled = false, ...restProps }: Props<T> = $props();
 
-    $inspect(group)
   const { input, label } = $derived(radio({ color, custom }));
 </script>
 
-<Label class={label({ class: labelClass })} color={disabled ? 'disabled' : undefined}>
+<Label class={label({ class: labelClass })} {disabled}>
   <input type="radio" bind:group {value} aria-describedby={aria_describedby} {disabled} aria-checked={group === value} class={input({ class: inputClass })} {...restProps} />
   {@render children()}
 </Label>